# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Advisor System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/{studentId}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student details
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
...
  /dashboard/student/{studentId}:
    get:
      operationId: getStudentDashboard
      tags: [dashboard]
      summary: Get student dashboard summary
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        year:
          type: integer
        gpa:
          type: number
        creditsCompleted:
          type: integer
        creditsTotal:
          type: integer
        advisorId:
          type: integer
          nullable: true
        planApproved:
          type: boolean
        planApprovedAt:
          type: string
          nullable: true
      required:
        - id
        - name
        - username
        - year
        - gpa
        - creditsCompleted
        - creditsTotal
        - planApproved
...
    CourseWithStatus:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        credits:
          type: integer
        year:
          type: integer
        category:
          type: string
        semesterOffered:
          type: string
        prerequisites:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [completed, in_progress, not_started]
        grade:
          type: string
          nullable: true
        prerequisitesMet:
          type: boolean
      required:
        - id
        - code
        - name
        - credits
        - year
        - category
        - semesterOffered
        - prerequisites
        - status
        - prerequisitesMet
...
    StudentDashboard:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        percentComplete:
          type: number
        completedCredits:
          type: integer
        remainingCredits:
          type: integer
        currentSemesterCourses:
          type: array
          items:
            $ref: "#/components/schemas/CourseWithStatus"
        upcomingDeadlines:
          type: array
          items:
            $ref: "#/components/schemas/Deadline"
        gpaHistory:
          type: array
          items:
            $ref: "#/components/schemas/GpaPoint"
      required:
        - student
        - percentComplete
        - completedCredits
        - remainingCredits
        - currentSemesterCourses
        - upcomingDeadlines
        - gpaHistory
    Deadline:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        type:
          type: string
          enum: [registration, advising, exam, other]
      required:
        - title
        - date
        - type
    GpaPoint:
      type: object
      properties:
        semester:
          type: string
        gpa:
          type: number
      required:
        - semester
        - gpa
...