# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Advisor System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  username: string;
  year: number;
  gpa: number;
  creditsCompleted: number;
  creditsTotal: number;
  advisorId?: number | null;
  planApproved: boolean;
  planApprovedAt?: string | null;
}