# lib/api-zod/src/generated/types/studentDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Advisor System API
 * OpenAPI spec version: 0.1.0
 */
import type { CourseWithStatus } from "./courseWithStatus";
import type { Deadline } from "./deadline";
import type { GpaPoint } from "./gpaPoint";
import type { Student } from "./student";

export interface StudentDashboard {
  student: Student;
  percentComplete: number;
  completedCredits: number;
  remainingCredits: number;
  currentSemesterCourses: CourseWithStatus[];
  upcomingDeadlines: Deadline[];
  gpaHistory: GpaPoint[];
}